requirements.txt
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
